monitoring/login-app/requirements.txt
httpx==0.27.0
fastapi==0.111.0
uvicorn==0.29.0
prometheus-client==0.20.0
opentelemetry-distro==0.45b0
opentelemetry-instrumentation-fastapi==0.45b0
opentelemetry-instrumentation-logging==0.45b0
opentelemetry-exporter-otlp==1.24.0
faker==18.11.1
prometheus_fastapi_instrumentator==7.0.0